Error checks.

+++ django/todo/tasks/views.py
@@ -1,5 +1,5 @@
 from django.shortcuts import render, redirect
-from django.http import HttpResponse
+from django.http import HttpResponse, Http404
 from tasks.models import Task
 import datetime
 # Create your views here.
@@ -11,26 +11,36 @@ def home_view(request):
 
 def add_view(request):
     return render(request, 'tasks/add.html')
+
 def edit_view(request):
     task_list = Task.objects.all()
     context_dict = {'tasks' : task_list}
-    return render(request, 'tasks/edit.html',context_dict)
+    return render(request, 'tasks/edit.html', context_dict)
+
 def impressum_view(request):
     return render(request, 'tasks/impressum.html')
+
 def add_todo(request):
-    text = request.POST['task']
-    date_tmp = request.POST['date'].split('/')
-    date = datetime.date(int(date_tmp[2]), int(date_tmp[0]), int(date_tmp[1]))
-    progress = int(request.POST['progress'])
-    if (progress < 0 or progress > 100) or (date - datetime.date.today()).days < 0:
-        return redirect('/add/')  
-    task = Task(task_text = text, task_progress = progress, task_deadline = date)
-    task.save()
-    return redirect('/')
+    if request.method == "POST":
+        text = request.POST['task']
+        date_tmp = request.POST['date'].split('/')
+        date = datetime.date(int(date_tmp[2]), int(date_tmp[0]), int(date_tmp[1]))
+        progress = int(request.POST['progress'])
+        if (progress < 0 or progress > 100) or (date - datetime.date.today()).days < 0:
+            return redirect('/add/')  
+        task = Task(task_text = text, task_progress = progress, task_deadline = date)
+        task.save()
+        return redirect('/')
+    return Http404("No.")    
   
 
 def del_todo(request):
-    id = request.POST['id']
-    to_delete = Task.objects.get(id = id)
-    to_delete.delete()
-    return redirect('/')
+    if request.method == "POST":
+        id = request.POST['id']
+        try:
+            to_delete = Task.objects.get(id = id)
+            to_delete.delete()
+        except Task.DoesNotExist:
+            raise Http404("Task does not exist")    
+        return redirect('/')
+    return Http404("No task to delete")
